atualizacao notificacao

+++ tech-challenge-fase2/api-notificacao/src/main/java/com/fase2/techchallenge/fiap/gestaonotificacao/dto/NotificacaoDTO.java
@@ -3,7 +3,7 @@ package com.fase2.techchallenge.fiap.gestaonotificacao.dto;
 import java.time.LocalDateTime;
 
 public record NotificacaoDTO(Long id,
-                             Long idVeiculo,
+                             Integer idVeiculo,
                              LocalDateTime dataHora,
                              String conteudo,
                              String situacao
